fix: remove mutex from cached session loading
This may improve the performance under load by limiting the size of the critical section. However, it’s only going to help if the cookie has been refreshed asynchronously in the background by a different process. If all workers share the same cookie (with same expiration), then wait is inevitable